backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Vector database
chromadb==0.4.18
# pinecone-client==2.2.4  # Alternative to ChromaDB

# AI and ML
openai==1.3.6  # For Llama API calls
sentence-transformers==2.2.2
numpy==1.24.3

# Audio processing
ffmpeg-python==0.2.0
pydub==0.25.1

# File monitoring
watchdog==3.0.0

# HTTP client
httpx==0.25.2
aiofiles==23.2.1

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
python-slugify==8.0.1
pillow==10.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Logging
structlog==23.2.0